+++ app/src/main/java/com/github/bumblebee202111/doubean/data/db/model/TopicEntity.kt
@@ -25,6 +25,9 @@ data class TopicEntity(
     @ColumnInfo("update_time")
     val updateTime: LocalDateTime,
 
+    @ColumnInfo("edit_time")
+    val editTime: LocalDateTime? = null,
+
     @ColumnInfo("like_count")
     val likeCount: Int? = null,