+++ CHANGELOG.md
@@ -3,7 +3,8 @@ Change Log
 
 ## Version 0.8.4(804)
 
-- Simplify group caching
+- Simplify caching of the group entity
+- Topic: Display edit time (if available) with animation
 - Fix Tab component min width for Group Getail
 
 ## Version 0.8.3(803)